🧹 Fix the if statement that prevents using conditions on Partition Keys (#6037)
This is a DynamoDB change that should have been in the previous PR, but failed to get committed.

It makes sure that in the one case where we have a table with only a PK, but no SK, the test that makes sure you only use `Condition`s on SKs works properly.

**How to test**

No need, it's been unit tested.