🧹 Introduce Python/HTML tests
Motivation
----------

Because it is hard to write Cypress tests, people often don't do it,
which increases the risk of regressions in the project. In this PR, I'm
making some changes to make it easier to write tests on the Flask code
in Python in the future, which should hopefully make it easier to
write tests overall. I'm thinking that Cypress tests have the following
downsides which lead us to not writing very many of them:

- Setup is awkward: to test a UI feature, you need the database in a
  situation where the tested UI element would be relevant. This means
  a program in state X, or a user in state Y. It is awkward to use the
  UI in Cypress to set up a database this way.
- Wait problems: complex interactions often have multiple steps, and it
  is easy to have race conditions in Cypress tests, where tests work
  on your machine but fail on the build server because of timing issues.
  Instead, in Python tests we'll make assertions on the HTML sent to
  the client, which is either good or bad in one shot.

Testing in Python is not a panacea: we will obviously not be able to
test anything that involves JavaScript, and we'll be asserting on
the HTML only which of  course tests nothing about presentation or
element visibility.

I'm not proposing to replace all our tests with Python/HTML tests, just
to get more coverage more easily this way.

If nothing else, the changes in this PR will help clean up the code
base a little 😉.

Motivation
----------

Previous, we used to create a global

```py
app = Flask(...)
```

object, and initialize it over the course of the `app.py` file.

This works well to get a server up and running, but it makes it harder
to test the code: for tests, you want fresh objects for every run.

In this PR, move the initialization code into a `create_app()` function,
so that we can create fresh instances of a `Flask` object on demand.
Also write some functions to clear the in-memory database state,
and write some Python/HTML tests to show what it looks like.