🧹 Add checks for duplicate HTML element IDs
In the course of investigating other issues, I've found that we are
emitting invalid HTML. For example, multiple elements with the same
`id="..."` attribute.

This doesn't cause loud errors in the browser, but interferes with the
proper functioning HTML selectors and Cypress tests don't always behave
like they should.

Add the start of a postprocessing step where we are going to validate
the HTML that is generated in Flask templates and crash loudly if it is
illegal.