🧹 Remove global variables from syntax highlighter
The syntax highlighter used to work by calling `initTranslations()` with
the current level and language, which then manipulates global variables
so that the syntax highlighter code has the right information to
highlight the different levels and languages of Hedy correctly.

To remove the need for global variables, we instead parameterize the
Lezer parsers and pass the level and language into the generator at
every time it is used.