🧹 Remove global variables from syntax highlighter (#6085)
The syntax highlighter used to work by calling `initTranslations()` with the current level and language, which then manipulates global variables so that the syntax highlighter code has the right information to highlight the different levels and languages of Hedy correctly.

To remove the need for global variables, we instead parameterize the Lezer parsers and pass the level and language into the generator at every time it is used.

The old parser code looked like this:

```ts
import {extendKeyword} from "./tokens"
export const parser = LRParser.deserialize({
  // ...

  specialized: [{term: 12, get: (value: any, stack: any) => (extendKeyword(value, stack) << 1) | 1, external: extendKeyword, extend: true}],
  //                                                         ^^^^^^ depends on language and level
});
```

`extendKeyword` depends on language and level, and so what we used to do is put the "current" language and level in a global variable and read the global variable in `extendKeyword`. This creates a sequencing requirement via global state, and apparently also a race condition because sometimes the syntax highlighters didn't load properly (see #6024).

In this PR, I change the generated source files to look like this:

```ts
import {specializeKeywordGen, extendKeywordGen} from "./tokens"
export function generateParser(level: number, language: string): LRParser {
  const extendKeyword = extendKeywordGen(level, language);
  // ^^^^This generates an 'extendKeyword' function that knows about leven and language

  return LRParser.deserialize({
    // ...
    specialized: [{term: 12, get: (value: any, stack: any) => (extendKeyword(value, stack) << 1) | 1, external: extendKeyword, extend: true}],
    //                                                         ^^^^ has the right level and language baked in already
})
}
```

This PR is a cleanup task from another larger PR I'm doing, that I'd like to separate out.

It also happens to fix #6024 😉 

**How to test**

This shouldn't change anything.