💻 Remove developer's mode in the new design (#6091)
First, some context is needed for this PR. It has been decided that to avoid a thousand paper cuts, we will release all big changes together later this year. We already have a /tryit page which is an altered copy of the existing /hedy page and allows us to receive feedback on the design. We will add all changes to this new /tryit page while keeping the current /hedy page as is. For all features that goes out of the /tryit page (e.g. removing the enforce-developer-mode in the class customization setting does not reside on the /tryit page), we will use a flag.

This PR:
- removes the developer mode from the /tryit page while preserving it on the /hedy page.
- removes the enforce-developer-mode class customization option behind a tryit flag.

Fixes #5987

**How to test**
Ensure that everything on the /hedy page works without changes:
- With developer mode disabled, you should be able to switch between parsons, quizzes and code adventures and the editor should be the correct height. The developer toggle should disappear when you are on a parsons or quizzes page. Make sure you test this with a parson problem with 5 lines (e.g. level 5, parson 2).
- With developer mode enabled, you should be able to switch back and forth between parsons and code adventures (using the buttons below the code output) and the editor should have the correct height. 
- When you go to the class customization page, you should be able to see the 'mandatory developer mode' checkbox.
- Setting its value should enforce the developer mode for the enrolled students.

Ensure the new features are working:
- The class customization page does not have the 'mandatory developer mode' option when you add `?tryit=yes` to the end of the url in your browser. Removing the suffix or changing it to `?tryit=no`, should make the option appear.
- The /tryit page should not display the developer toggle. The height of the editor should be correct when switching between adventures and quizzes. Please do not test the parsons on the /tryit page because they will be removed next and it is ok if they are temporarily not working as expected.
- When a class has the developer mode enforced, the /tryit page should not hide the explanations.