💻 Remove default level of new adventures (#6095)
This PR removes the pre-selected level 1 when a new adventure is created.

Fixes #5922

**How to test**
Go to the for-teachers page and create a new adventure. When the form appears, check that no levels are pre-selected. Check that you cannot save an adventure without setting levels. Now try updating an existing adventure and again check that you cannot save an adventure if there isn't at least 1 level selected.