Revert "📜 Add tags to mailchimp users (#6080)" (#6120)
This reverts commit 4e1260388caca5110c5cc1880f6c612e1c7f5cf0.

Temporarily reverts the change in #6080 

The revert is needed because the content for the Mailchimp emails is not ready. Releasing this change without the counterpart in Mailchimp, would result in either users receiving irrelevant content or double emails. We will merge this back to mail when the content is ready.