🧹 Remove warnings during tests
We were persistently printing 2 deprecation warnings during unit tests.
These warning originated from old versions of the Lark parser and the
Pydantic library, that were using deprecated Python APIs.

Upgrade both libraries to newer versions; the warnings are now gone.

Because we cache Lark parsers on disk, we need to make sure to mix the
Lark version into the cache key so that we don't load old parsers into a
new version of the library.