💻 Replace public adventures pagination with infinite scroll (#6124)
The public adventures page had a pagination mechanism, with "next page"/"prev page" buttons.

A more modern idiom these days is to use "infinite scroll": once the user scrolls to the bottom of the container, we fetch the next bit of results, and so on.

**How to test**

Go to the public adventures page, scroll to the bottom of the result list, try to catch a glimpse of the spinner before the next set of results loads.

<img width="345" alt="image" src="https://github.com/user-attachments/assets/dd98b55b-7632-4de1-8093-99b67a363f2b" />