💻 Show the reason why you can see adventures (#6134)
You can see adventures of teachers that you share classes with: both  if you are the second teacher, and if they are the second teacher.

This is surprising at first, so this adds an icon that indicates why you can see a given adventure.

<img width="844" alt="image" src="https://github.com/user-attachments/assets/01ddcd93-f8c8-4101-9177-6fcce0c33dcc" />


**How to test**

Log in as a user that shares a class with another teacher. Observe that you can see their adventures.

Closes #6133.