🪲 Set highlighter language and level  (#6126)
PR #6085 makes uses of `data-level` and `data-kwlang` but these are never set. As a result, the language of the highlighter always defaults to English regardless of the current language. This PR makes some minimal changes to address the issue.

Fixes #6112

**How to test**
As an anonymous user go to /hedy and change the language to Turkish (or any other language). In the editor paste `yazdır test` and check that the first word is highlighted as a command. Do the same check for the /tryit page and a logged in user.



Co-Authored-By: Rico Hermans <rix0rrr@gmail.com>
Co-Authored-By: rix0rrr <rix0rrr@users.noreply.github.com>
Co-Authored-By: Rico Huijbers <rix0rrr@gmail.com>