🪲 Fix race condition on customization's update (#6217)
The problem in the issue, as Boryana pointed out, was a race condition caused by making two calls at roughly the same time:  to /update-customizations, and to /get-customization-level. This last call also performs some updates in the customization object, so if this one is done first, the customization object only gets the old update.

To address this I first tried just not triggering that second update, but then a few tests failed. So my new approach is the following:

1. I deleted the parameter from the autosave that triggers that second call. It doesn't really make sense to make that call if we are not sure that the first update was completed, so I moved it to the `save_customizations` function.
2. I modified the `get_class_info` to only update the customizations in case of a first load. If we are accessing this function through HTML calls, it should not be doing updates.


Fixes #6176 

**How to test**

Follow these steps to verify this PR works as intended:

* Updating the other options in customizations should work.