💻 Send notification when a teacher requested pairing  (#6222)
This PR introduces email notifications when a teacher signs up and requests to be paired with another teacher.

Fixes #6211

**How to test**
When you sign up as a teacher and tick the "I would like to be paired with another teacher for help" checkbox, you should receive an email. However, emails are not sent when running Hedy locally.

If you would like to actually test receiving an email and you do not have an aws account, please ping me so that I can provide you with credentials. Alternatively, you can debug the code and check that the sending email.