🪲 Add link to class information when rerendering page (#6236)
After checking a checkbox in adventure overview table, trying to copy the invite link wouldn't work because the link wasn't include in the class information sent to the template.

I know we are going to remove this feature soon, but since this is a very small change I figured we could fix it in the interim before that happens.

**How to test**

Follow these steps to verify this PR works as intended:

* Go to class 1 and check one of the student's adventures as reviewed.
* Click on the "copy invite link" button. 
* Access the link and it should work.