💻 Hide copy class link button from redesign (#6238)
Hides the button to get the class invite link when the `is_redesign_enabled` flag is set to true. Since I can't delete the actual functions without breaking the current functionality, I marked them with a TODO so we know to delete them once the redesign is deployed.

We also need to delete the index on the link field of the class table, which can be done without disrupting the table, but the actual fields would need to be deleted manually from the DB.

Fixes  #5977 

**How to test**

* Set the redesign flag to true
* go to a class page and see that the Copy join link button is hidden.
* If the flag is false, the button should work as normally.