🪲 fixes 6215 - introduces a new password generator that doesn't use characters that are easy to misread (#6232)
We use a new random_password_generator function that uses a subset (base58) of the available characters to generate passwords. The reason for this change is that some characters are easy to misread, and this can lead to confusion.

**How to test**

Follow these steps to verify this PR works as intended:

* Invite new students to a group
* Print the generated passwords
* Confirm that none of the passwords contain the characters 0, O, l, and I

**Checklist**
Done? Check if you have it all in place using this list: (mark with x if done)

- [x] Contains one of the PR categories in the name
- [x] Describes changes in the format above
- [x] Links to an existing issue or discussion
- [x] Has a "How to test" section