💻 Put minimum quiz score behind a flag (#6242)
Removes the effect of the minimum quiz score set up by a teacher when the redesign flag is enabled.

Fixes #5989

**How to test**
Check that when the utils.is_redesign_enabled is set to True a student can see all levels regarding of whether they have completed a quiz for the previous level or not. Currently, in the local DB seeding, student1 cannot access levels above 13 and can be used for this test.