🪲 Fix password autogeneration (#6266)
This PR fixes incorrect auto-generated passwords when a teacher creates student accounts in bulk.

Fixes #6267 

**How to test**
Log in as a teacher and create student accounts in bulk without providing custom passwords. When the accounts are created (and their respective passwords are generated), the passwords should be 6 characters each and they should be different from one another.