💻 Add adventure solutions (#6264)
With this PR teachers can view the solutions of adventures if such solutions exist.

Fixes # 6261

**How to test**
Currently, no content exists for solutions, so to test this, you will need to add solutions. For example, add the following snippets to the content/adventures/en.yaml file:
```
  answer_code: |
      ```
      print this is the solution code 
      ```
  answer_note: |
      this is an explanation of the solution
  answer_code_2: |
      ```
      print this is the solution code 2
      ```
  answer_note_2: |
      this is an explanation of the solution 2
  answer_code_3: |
      ```
      print this is the solution code 3
      ```
```
Note that `answer_code*` is a required parameter, but `answer_note*` is not. Once the data is there, go to the /tryit page and test at least the following scenarios:
- [ ] teachers and second teachers should see the solution tabs. Students should not see the tabs.
- [ ] switching between the Adventure and Solution tabs of an adventure
- [ ] switching between adventures that have solutions and such that do not have solutions