Address comments on #6032 (#6291)
Addresses @rix0rrr comments in the previous PR, except for calling HTMX from TypeScript. It's not that I'm opposed to the change, or don't think the comment raises a fair concern, but we are in a tight deadline and that change might take a little more time. @boryanagoncharenko if that's a change you think it's worth doing, though, I can invest some time in the following days to change the logic in the way the search operates.

**How to test**
The autocomplete should work!


Co-Authored-By: Boryana Goncharenko <3010723+boryanagoncharenko@users.noreply.github.com>
Co-Authored-By: boryanagoncharenko <boryanagoncharenko@users.noreply.github.com>