📜 Use translate string of colors and random (#6326)
Noticed that some colors used in context of `{color}` do not use the string with `{}`. I tried to fix this by replacing e.g. `{color} red` with `{color} {red}`.

Also noticed that `random` is used without `{}`, also tried to fix that.

**How to test**
Test the examples with the English colors words and random word compared to use the translated words in your native language. 

**Checklist**

- [x] Contains one of the PR categories in the name
- [x] Describes changes in the format above
- [ ] Links to an existing issue or discussion
- [x] Has a "How to test" section