🪲 Teachers cannot be invited as second teachers (part 1) (#6331)
This PR is a part of 2-step fix. It changes the value of the `is_teacher` attribute for newly created teachers from `true` to `1`. Correcting the wrong attribute values for existing users will be done separately of this PR through a one-time-script.

As a side effect, it addresses the following problem: a teacher with `is_teacher=true` cannot be invited as a second teacher because they cannot appear in the invitation modal.

Fixes #6321

**How to test**
This issue cannot be replicated locally - it requires dynamo db. Once this is deployed to alpha, create a new teacher and try inviting them as a second teacher of a class. Their username should appear in the search pop up.