🧹 Migrate to `logger.warning` (#6349)
## PR Summary
This small PR migrates from the deprecated `logger.warn` to the recommended `logger.warning` method.

**How to test**

Follow these steps to verify this PR works as intended:

**Checklist**
Done? Check if you have it all in place using this list: (mark with x if done)

- [x] Contains one of the PR categories in the name
- [x] Describes changes in the format above
- [ ] Links to an existing issue or discussion
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!