🧹 Path from `ROOT_DIR` instead of `__file__/..`
"Pathing up" via `__file__` (e.g. `../`) is fragile, so only do it in a
centrally-defined `config.ROOT_DIR`, and path from root everywhere else

Also fix improper Python exec-as-module vs exec-as-script in `dodo.py`.
Needing to modify `sys.path` in order to import is a sign of having used
the wrong variant. For readers, a rule of thumb:
 * `python foo/bar.py` - independent, only depends on contents of `foo`
 * `python -m foo.bar` - depends on project modules outside of `foo`

These changes also facilitate a future migration towards the cleaner
[src layout](https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/)

In the future, this can be enforced with custom pylint checks, which is
easier to add after migrating the project to use `pyproject.toml`
instead of `requirements.txt`