🧹 Stop calling HTMX from JS (#6351)
I want to admit that I was wrong and doing this was actually easier than I thought 😅 

I generalized the htmx search modal a little bit, so that client code can send the htmx attributes they want to set before doing the requests. This still uses JS, of course, but no JS is used to actually make the calls, just to set things up, the alternative to this would be to stop using our `modal` object and set things up when rendering the page from Jinja.

The functionality should remain the same, as this PR only refactors things a bit.

Fixes #6293

**How to test**

* Log in as a teacher and invite a few students and second teachers.


Co-Authored-By: boryanagoncharenko <boryanagoncharenko@users.noreply.github.com>