🧹 Clean code related to already removed features (#6354)
Now that we were able to ship our new design for the adventures page, we can begin to remove from the code all of the features that we removed in version, most notably quizzes and parsons.

So, a summary:

* It removes some of the html files that weren't needed. Some of them I kept, to maintain two endpoints called /adventures and /render-code. These two will be fixed when we get onto refactoring the teacher's page.
* Removes all of the content for the quizzes and parsons.
* Removes all mentions to quizzes and parsons in the code, except for the ones in the section for teachers, that one will be refactored shortly.
* It moves the hour-of-code endpoint to the new design. In this vein, I also removed the capability to pass a subset of adventures through the URL, something that was not really used and is fixed through customizations.

This changes shouldn't affect the way the site works, since all of these was unused code! Still I'd like if someone could do a little round of testing around and check that things are indeed working :)

Fixes #6318

**How to test**

* Everything should be working as expected