feat: add a script to print a language's workbook to a PDF
This adds a page that prints all the workbooks of a single language one
after the other, and a script that uses Puppeteer to visit that page and
export the page to a PDF.

This makes it possible to generate a printable PDF for any workbook
with little effort.

The script currently works against a local server, so you need to
run a copy of the devserver on your machine before running the script.